+++ packages/_shared/src/index.tsx
@@ -32,4 +32,8 @@ export { entityHelpers };
 import * as ConstraintsUtils from './utils/constraints';
 export { ConstraintsUtils };
 
+export { isValidImage } from './utils/isValidImage';
+
+export { shortenStorageUnit, toLocaleString } from './utils/shortenStorageUnit';
+
 export { Entry, File, Asset } from './typesEntity';