+++ packages/validation-errors/src/ValidationErrors.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { ValidationError, Link } from '@contentful/app-sdk';
 import type {
   SpaceAPI,
   Entry,
@@ -13,12 +14,6 @@ import * as styles from './styles';
 import { TextLink, List, ListItem } from '@contentful/f36-components';
 
 import { ExternalLinkIcon, InfoCircleIcon } from '@contentful/f36-icons';
-
-export type ValidationError = {
-  name: string;
-  message: string;
-  conflicting?: Array<{ sys: { id: string } }>;
-};
 
 type UniquenessErrorProps = {
   error: ValidationError;
@@ -62,9 +57,13 @@ function UniquenessError(props: UniquenessErrorProps) {
     [props.localeCode, props.defaultLocaleCode, contentTypesById]
   );
 
+  let conflicting: Link<'Entry', 'Link'>[] = [];
+  if ('conflicting' in props.error) {
+    conflicting = props.error.conflicting;
+  }
   React.useEffect(() => {
     const entryIds = state.entries.map((entry) => entry.id);
-    const conflictIds = props.error.conflicting?.map((entry) => entry.sys.id) || [];
+    const conflictIds = conflicting.map((entry) => entry.sys.id);
 
     // Avoid unnecessary refetching
     if (conflictIds.every((id) => entryIds.includes(id))) {
@@ -89,7 +88,7 @@ function UniquenessError(props: UniquenessErrorProps) {
         entries,
       });
     });
-  }, [getTitle, state.entries, props.error.conflicting, props.space.getEntries, props.getEntryURL]);
+  }, [getTitle, state.entries, conflicting, props.space.getEntries, props.getEntryURL]);
 
   return (
     <List className={styles.errorList} testId="validation-errors-uniqueness">