+++ packages/default-field-editors/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.2](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@0.10.1...@contentful/default-field-editors@0.10.2) (2021-10-18)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+
+
+
+
+## [0.10.1](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@0.10.0...@contentful/default-field-editors@0.10.1) (2021-10-14)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+
+
+
+
+# [0.10.0](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@0.9.17...@contentful/default-field-editors@0.10.0) (2021-10-06)
+
+
+### Features
+
+* [EXT-3161] bump app sdk to v4 ([#881](https://github.com/contentful/field-editors/issues/881)) ([9c4a2af](https://github.com/contentful/field-editors/commit/9c4a2af07da203d59fb5f15c3a5188ecc64b1d44))
+
+
+
+
+
+## [0.9.17](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@0.9.16...@contentful/default-field-editors@0.9.17) (2021-10-06)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+
+
+
+
+## [0.9.16](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@0.9.15...@contentful/default-field-editors@0.9.16) (2021-09-20)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+
+
+
+
+## [0.9.15](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@0.9.14...@contentful/default-field-editors@0.9.15) (2021-09-17)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+
+
+
+
+## [0.9.14](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@0.9.13...@contentful/default-field-editors@0.9.14) (2021-09-17)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+
+
+
+
+## [0.9.13](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@0.9.12...@contentful/default-field-editors@0.9.13) (2021-09-16)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+
+
+
+
 ## [0.9.12](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@0.9.11...@contentful/default-field-editors@0.9.12) (2021-08-26)